[worker] extract redundant code for reading shard from cli.
